fixed some minor typos (#17)
Fixed minor typos. Also noticed the link to 'CONTRIBUTING.md' was
slightly off - idk if that was intentional. I noticed some links don't
go anywhere but they were placed preemptively. I wasn't sure if that was
the case or not for 'CONTRIBUTING.md' but seeing as that was an actual
file in the repo I updated the link to send visitors there.